[Refactor] (BSP/Core): 重构外设和日志系统初始化
将LED和Logger外设实例从BSP层移动到Core层的peripherals模块
移除BSP层中冗余的全局变量定义和实现文件
统一日志系统的USART定义和任务管理
添加缺失的#pragma once指令到多个头文件

这些改动提高了代码的组织性和模块化程度，使得外设管理更加集中
