+++ Source/BSP/inc/bsp_led.hpp
@@ -11,5 +11,3 @@ class LED {
    private:
     GPIO gpio;
 };
-
-extern LED sysLed;
