--- Source/BSP/src/bsp_led.cpp
@@ -1,3 +1,0 @@
-#include "bsp_led.hpp"
-
-LED sysLed(GPIO::Port::C, GPIO::Pin::PIN_13);
