+++ Source/Core/CMakeLists.txt
@@ -2,8 +2,9 @@ set(COMMON_SOURCES ./gd32f4xx_it.c ./main.cpp ./protocal.cpp)
 add_subdirectory(uwb)
 if(SLAVE)
   target_sources(
-    ${EXECUTABLE_NAME} PRIVATE ${COMMON_SOURCES} ./slave/slave_mode.cpp
-                               ./slave/harness.cpp ./slave/msg_proc.cpp)
+    ${EXECUTABLE_NAME}
+    PRIVATE ${COMMON_SOURCES} ./slave/slave_mode.cpp ./slave/harness.cpp
+            ./slave/msg_proc.cpp ./slave/peripherals.cpp)
   target_include_directories(${EXECUTABLE_NAME} PUBLIC . ./slave)
 else()
   target_sources(
