+++ Source/Core/slave/msg_proc.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <vector>
 
 #include "FreeRTOS.h"
