+++ Source/Core/slave/res.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <cstdint>
 
 #include "bsp_adc.hpp"
