+++ Source/Core/slave/slave_mode.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <array>
 #include <cmath>
 #include <cstddef>
